fix: fix for custom-copilot-basic ai chat bot template (#11043)
* fix: combine py files and some fix

* fix: bad template name

* fix: naming
